staff-service/requirements.txt
django>=4.2
pymysql
cryptography
djangorestframework
requests
